refactor: list components #1
