+++ src/views/ScanQr.vue
@@ -1,10 +1,14 @@
 <script setup>
+import { useStore } from 'vuex'
+
 import GoBack from '@/components/go-back.vue'
 import CodeScanner from '@/components/code-scanner.vue'
 import OptimizeRouteButton from '@/components/optimize-route-button.vue'
 import CapacityInput from '@/components/capacity-input.vue'
-import InputPointsList from '@/components/input-points-list.vue'
+import ListContainer from '@/components/list-container.vue'
+import InputPointItem from '@/components/input-point-item.vue'
 
+const store = useStore()
 </script>
 
 <template>
@@ -14,7 +18,13 @@ import InputPointsList from '@/components/input-points-list.vue'
     <h1 class="text-center text-2xl font-bold my-4">Scan QR Code</h1>
     <code-scanner />
 
-    <input-points-list />
+    <list-container title="Added input points">
+      <li v-for="point in store.state.inputPoints" :key="point.id"
+          class="p-4 rounded-lg border-2 border-gray-200 hover:bg-gray-200 transition-colors duration-300"
+      >
+        <input-point-item :point="point" />
+      </li>
+    </list-container>
 
     <div>
       <h1 class="text-center text-2xl font-bold my-4">Input vehicles capacity</h1>
